+++ CHANGELOG.md
@@ -26,6 +26,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dispatch Mappings
 - Flows
 - Logi reportings
+- Business Hours
+- API Keys
+- Message Templates
+
+## [4.79.2] - 2019-09-13
+### Added
+- <no-jira> - Adding Business Hours, Api Keys and Message Templates page to beta features.
 
 ## [4.79.1] - 2019-09-11
 ### Added